Use the page- rather than target-related asset variable.
The benefit is that when you have pages from more than one folder
level in your list of files, `options.assets` cannot be set
correctly for each file. By looking in the `options.pages` variable
we can get the page-related asset variable which leads to better paths.
The code falls back to the `option.asset` variable if for some reason
a page asset variable could not be found.